refactor: one shared header for the sync session wire contract
The WZSY frame layout, its bounds and the session-id charset were
defined twice: once in the intake (syncFrame/sessionId) and once,
hand-copied, in the producer's socket transport. A wire change would
have to hit both and nothing would tell you if it missed one.

The contract now lives in sync_session_wire.hpp under sync_protocol,
the layer both ends already reach. The intake keeps its framing code
and takes the constants and isValidSessionId() from there; sessionId.hpp
had nothing left and is gone.
