+++ src/client-agent/https_client/tests/unit/syncFrame_test.cpp
@@ -14,7 +14,7 @@
 
 #include "syncFrame.hpp"
 
-#include "sessionId.hpp"
+#include "shared_modules/sync_protocol/include/sync_session_wire.hpp"
 
 #include <gtest/gtest.h>
